Move kotlin.internal.jdk8.JDK8PlatformImplementations to kotlin-stdlib.
