+++ metadata/index.json
@@ -106,6 +106,10 @@
     "module": "com.github.ben-manes.caffeine:caffeine"
   },
   {
+    "directory": "org.jetbrains.kotlin/kotlin-stdlib",
+    "module": "org.jetbrains.kotlin:kotlin-stdlib"
+  },
+  {
     "directory": "org.jetbrains.kotlin/kotlin-reflect",
     "module": "org.jetbrains.kotlin:kotlin-reflect"
   },
