+++ metadata/org.jetbrains.kotlin/kotlin-reflect/1.7.10/reflect-config.json
@@ -89,17 +89,5 @@
     },
     "name": "kotlin.reflect.jvm.internal.impl.resolve.scopes.DescriptorKindFilter",
     "allPublicFields": true
-  },
-  {
-    "condition": {
-      "typeReachable": "kotlin.internal.jdk8.JDK8PlatformImplementations"
-    },
-    "name": "kotlin.internal.jdk8.JDK8PlatformImplementations",
-    "methods": [
-      {
-        "name": "<init>",
-        "parameterTypes": []
-      }
-    ]
   }
 ]
