+++ metadata/org.jetbrains.kotlin/kotlin-stdlib/1.7.10/index.json
@@ -1,0 +1,3 @@
+[
+  "reflect-config.json"
+]
