+++ metadata/org.jetbrains.kotlin/kotlin-stdlib/1.7.10/reflect-config.json
@@ -1,0 +1,14 @@
+[
+  {
+    "condition": {
+      "typeReachable": "kotlin.internal.jdk8.JDK8PlatformImplementations"
+    },
+    "name": "kotlin.internal.jdk8.JDK8PlatformImplementations",
+    "methods": [
+      {
+        "name": "<init>",
+        "parameterTypes": []
+      }
+    ]
+  }
+]
