+++ metadata/org.jetbrains.kotlin/kotlin-stdlib/index.json
@@ -1,0 +1,10 @@
+[
+  {
+    "latest": true,
+    "metadata-version": "1.7.10",
+    "module": "org.jetbrains.kotlin:kotlin-stdlib",
+    "tested-versions": [
+      "1.7.10"
+    ]
+  }
+]
