+++ tests/src/index.json
@@ -264,6 +264,17 @@
     ]
   },
   {
+    "test-project-path": "org.jetbrains.kotlin/kotlin-stdlib/1.7.10",
+    "libraries": [
+      {
+        "name": "org.jetbrains.kotlin:kotlin-stdlib",
+        "versions": [
+          "1.7.10"
+        ]
+      }
+    ]
+  },
+  {
     "test-project-path": "org.jetbrains.kotlin/kotlin-reflect/1.7.10",
     "libraries": [
       {
