+++ tests/src/org.jetbrains.kotlin/kotlin-stdlib/1.7.10/build.gradle
@@ -12,8 +12,11 @@ plugins {
 String libraryVersion = tck.testedLibraryVersion.get()
 
 dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-reflect:$libraryVersion"
-    testImplementation 'org.assertj:assertj-core:3.22.0'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$libraryVersion"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.6.4"
+
+    implementation 'org.jetbrains.kotlin:kotlin-test-junit:1.7.10'
+
 }
 
 compileTestKotlin {
