+++ tests/src/org.jetbrains.kotlin/kotlin-stdlib/1.7.10/gradle.properties
@@ -1,0 +1,2 @@
+library.version = 1.7.10
+metadata.dir = org.jetbrains.kotlin/kotlin-stdlib/1.7.10
